projekt/car-racing-rl/requirements.txt
gymnasium
numpy
tensorflow
matplotlib
pandas
opencv-python
scikit-learn
neat-python